+++ src/main/java/wercsmik/spaghetticodingclub/domain/auth/service/EmailService.java
@@ -12,7 +12,7 @@ import org.springframework.mail.javamail.MimeMessageHelper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wercsmik.spaghetticodingclub.domain.auth.entity.EmailVerification;
-import wercsmik.spaghetticodingclub.domain.auth.repotiroy.EmailVerificationRepository;
+import wercsmik.spaghetticodingclub.domain.auth.repository.EmailVerificationRepository;
 import wercsmik.spaghetticodingclub.global.exception.CustomException;
 import wercsmik.spaghetticodingclub.global.exception.ErrorCode;
