+++ src/main/java/wercsmik/spaghetticodingclub/domain/auth/service/EmailVerificationCleanupService.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
-import wercsmik.spaghetticodingclub.domain.auth.repotiroy.EmailVerificationRepository;
+import wercsmik.spaghetticodingclub.domain.auth.repository.EmailVerificationRepository;
 
 @Service
 @RequiredArgsConstructor
